backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.7
spacy==3.7.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0